Refactor and update filenames
