+++ bot/src/index.ts
@@ -1,7 +1,7 @@
 import { setTimeout as delay } from 'timers/promises';
 import { setupRpcUrlAndGetNetworks } from 'auctions-core/src/rpc';
-import { loopCollateral } from './auctions';
-import { loopSurplus } from './surplus';
+import { loopCollateral } from './auctions/collateral';
+import { loopSurplus } from './auctions/surplus';
 import { setupKeeper } from './keeper';
 import { RPC_URL } from './variables';
 import { setupTwitter } from './twitter';
